requirements.txt
Flask==0.12.2
requests==2.18.3
PyYAML==3.12
